Refactored diagram_filter initialization to a factory (#289)
