+++ src/common/generators/json/generator.h
@@ -18,7 +18,7 @@
 #pragma once
 
 #include "common/generators/generator.h"
-#include "common/model/diagram_filter.h"
+#include "common/model/filters/diagram_filter.h"
 #include "config/config.h"
 #include "util/error.h"
 #include "util/util.h"
